Time: 0 ms (100.00%) | Memory: 7.7 MB (65.24%) - LeetSync

+++ 371-sum-of-two-integers/sum-of-two-integers.cpp
@@ -5,7 +5,7 @@ public:
         while(b != 0) {
 
             // carry
-            unsigned carry = (unsigned)(a & b) << 1;
+            int carry = (a & b) << 1;
 
             // sum without carry
             a = a ^ b;
